use the pre-existing test helper for this. Add parse method to helper
